TS: Splitting DomainsOnProtein
